refactor(cmd): move authcmd to cmd/ and refactor flags in bridge,light,and full in cmd/celestia
